chore: rename locales and constants
