+++ src/app/ui/footer/lib/constants.ts
