+++ src/app/ui/footer/locales/index.ts
