+++ src/pages/landing/locales/index.ts
